refactor(find-wallet): fold filters header into WalletFilters
WalletFiltersHeader had a single caller and reached back into its sibling
for the four filter-key constants. Folding it into WalletFilters lets the
reset path loop ALL_FILTER_KEYS instead of four hand-written setFilter
calls, and reuse the existing trackFilterToggle helper.

Also normalize WalletPersonaId to import from data/wallets/personas
rather than the walletData re-export, matching the sibling components.
